Revert "Code reorganization parties.modelys.py"
