+++ notifications/models.py
@@ -1,7 +1,5 @@
 from django.db import models
 from django.conf import settings
-
-from parties.models import Party
 # Create your models here.
 
 class NotificationManager(models.Manager):
@@ -14,10 +12,9 @@ class NotificationManager(models.Manager):
 class Notification(models.Model):
 	action 				= models.CharField(max_length=30)
 	user 				= models.ForeignKey(settings.AUTH_USER_MODEL,
-						on_delete=models.CASCADE,related_name="notifs_list")
-	time_created 		= models.DateTimeField(auto_now_add=True)
-	party 				= models.ForeignKey(Party,
 						on_delete=models.CASCADE)
+	time_created 		= models.DateTimeField(auto_now_add=True)
+	party 				= models.IntegerField(default = 0)
 	seen 				= models.BooleanField(default=False)
 
 	# this just allows the notification objects to use the manager methods
