+++ parties/tasks.py
@@ -6,7 +6,6 @@ from celery import shared_task
 from userstatistics import statisticsfunctions
 from .models import Party
 from notifications.models import Notification
-from parties import partyHandling
 # the shared task just makes it so the celery app can access this
 @shared_task
 # this method takes the list of joined, reorders it randomly, and picks one
@@ -20,7 +19,7 @@ def pick_winner(party_id):
 		return 
 	# for any party that hasnt closed by end time, tell the owner its closing
 	if party.is_open:
-		Notification.objects.create(user=party.user, party=party,\
+		Notification.objects.create(user=party.user, party=party.pk,\
 		action="owner_event_close")
 	# if there are people that joined the event
 	if party.joined.all().count() > 0:
@@ -32,7 +31,7 @@ def pick_winner(party_id):
 			for i in range(0,party.num_possible_winners):
 				if pool:
 					winner = pool.first()
-					partyHandling.win_toggle(winner, party)
+					Party.objects.win_toggle(winner, party)
 					pool = pool.exclude(pk=winner.pk)
 			statisticsfunctions.lottery_update_end_stats(party)
 		#If the party event is a bid and hasnt closed for some reason
@@ -43,7 +42,7 @@ def pick_winner(party_id):
 				print (w)
 			#add winners in
 			for i in winners:
-				partyHandling.win_toggle(i, party)
+				Party.objects.win_toggle(i, party)
 			statisticsfunctions.bid_update_end_stats(party)
 		elif party.event_type==3 and party.is_open:
 			print("Buyout event is over")	
@@ -68,7 +67,7 @@ def pick_winner(party_id):
 	if party.winners.all().count() > 0:
 		notification_list = party.winners.all()
 		for n in notification_list:
-			Notification.objects.create(user=n, party=party,\
+			Notification.objects.create(user=n, party=party.pk,\
 			action="fan_reminder")
-	Notification.objects.create(user=party.user, party=party,\
+	Notification.objects.create(user=party.user, party=party.pk,\
 	action="owner_reminder")
